requirements.txt
flask
python-dotenv
flask-mysql
cryptography
flask-jwt-extended
colorama
bcrypt
flask-socketio
flask-cors
python-socketio
eventlet